LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-video-element/intrinsic_sizes-expected.txt

spec reference



PASS default object size is 300x150
FAIL default height is still 150 assert_equals: expected "150px" but got "200px"
FAIL default width is still 300 assert_equals: expected "300px" but got "200px"
PASS default object size after src is removed
PASS default object size after poster is removed
